# backend/src/modules/categories/categories.controller.ts
import { Body, Controller, Get, Param, ParseUUIDPipe, Patch, Post, UseGuards } from "@nestjs/common";

import { GetCurrentUser } from "../auth/decorators/get-current-user.decorator";
import { Roles } from "../auth/decorators/roles.decorator";
import { JwtAuthGuard } from "../auth/guards/jwt-auth.guard";
import { RolesGuard } from "../auth/guards/roles.guard";
import { User, UserRole } from "../users/entities/user.entity";
import { CategoriesService } from "./categories.service";
import { CreateCategoryDto, UpdateCategoryDto } from "./dto/category.dto";

/** Danh mục đang nhận chiến dịch (công khai). */
@Controller("categories")
export class CategoriesController {
  constructor(private readonly categoriesService: CategoriesService) {}

  @Get()
  list() {
    return this.categoriesService.listActive();
  }
}

/** Quản trị danh mục: thêm, đổi tên (cập nhật cả chiến dịch), sắp xếp, tắt/bật. */
@Controller("admin/categories")
@UseGuards(JwtAuthGuard, RolesGuard)
@Roles(UserRole.ADMIN)
export class AdminCategoriesController {
  constructor(private readonly categoriesService: CategoriesService) {}

  @Get()
  list() {
    return this.categoriesService.listAll();
  }

  @Post()
  create(@Body() dto: CreateCategoryDto, @GetCurrentUser() admin: User) {
    return this.categoriesService.create(dto, admin);
  }

  @Patch(":id")
  update(@Param("id", ParseUUIDPipe) id: string, @Body() dto: UpdateCategoryDto, @GetCurrentUser() admin: User) {
    return this.categoriesService.update(id, dto, admin);
  }
}