# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Crowdfunding Platform API
  version: 0.3.0
  description: API cho nền tảng gây quỹ cộng đồng Góp Mầm.
servers:
  - url: http://localhost:4000/api/v1
    description: Backend local
paths:
...
  /campaigns:
    get:
      operationId: listCampaigns
      summary: Danh sách chiến dịch công khai
      description: >-
        Không cần đăng nhập. Mặc định chỉ trả chiến dịch ở trạng thái công
        khai (`approved`, `active`, `success`, `ended`) trừ khi lọc `status`.
        Lọc `status` chỉ nhận trạng thái đã công khai; draft/pending/needs_info/rejected/cancelled trả 400.
        Mọi bộ lọc chạy ở server (truy vấn lấy id rồi mới nạp quan hệ nên phân trang đúng).
        `q`/`location` thoát ký tự `%`/`_` của LIKE. `sort=progress` theo tỷ lệ
        `currentAmount/goalAmount`; `sort=ending` bỏ chiến dịch đã quá hạn. Khoảng ngược
        (`minGoal > maxGoal`, `minProgress > maxProgress`) trả 400.
      parameters:
        - $ref: "#/components/parameters/CampaignQ"
        - $ref: "#/components/parameters/CampaignCategory"
        - $ref: "#/components/parameters/CampaignStatusFilter"
        - name: location
          in: query
          description: Địa điểm chứa chuỗi này
          schema:
            type: string
            maxLength: 100
        - name: minGoal
          in: query
          description: Mục tiêu vốn tối thiểu (VNĐ)
          schema:
            type: number
            minimum: 0
        - name: maxGoal
          in: query
          description: Mục tiêu vốn tối đa (VNĐ)
          schema:
            type: number
            minimum: 0
        - name: minProgress
          in: query
          description: Tỷ lệ hoàn thành tối thiểu (%)
          schema:
            type: number
            minimum: 0
            maximum: 1000
        - name: maxProgress
          in: query
          description: Tỷ lệ hoàn thành tối đa (%)
          schema:
            type: number
            minimum: 0
            maximum: 1000
        - name: endingWithinDays
          in: query
          description: Chỉ chiến dịch `active` kết thúc trong N ngày tới
          schema:
            type: integer
            minimum: 1
            maximum: 365
        - $ref: "#/components/parameters/CampaignSort"
        - $ref: "#/components/parameters/Offset"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Danh sách chiến dịch có phân trang
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CampaignList"
        "400":
          description: Trạng thái chưa công khai, tham số sai kiểu/ngoài phạm vi hoặc khoảng lọc ngược
    post:
      operationId: createCampaign
      summary: Tạo chiến dịch mới (trạng thái `draft`)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCampaignRequest"
      responses:
        "201":
          description: Đã tạo chiến dịch
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
  /campaigns/mine:
    get:
      operationId: listMyCampaigns
      summary: Danh sách chiến dịch của chính chủ dự án đang đăng nhập
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/CampaignStatusFilter"
        - $ref: "#/components/parameters/Offset"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Danh sách chiến dịch của tôi (mọi trạng thái)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CampaignList"
  /campaigns/{id}:
    get:
      operationId: getCampaignById
      summary: Chi tiết chiến dịch
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: Thông tin chiến dịch
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updateCampaign
      summary: Cập nhật chiến dịch
      description: >-
        Chỉ chủ dự án hoặc admin. Chỉ sửa được khi ở `draft`, `rejected` hoặc
        `needs_info`. Ghi audit `campaign.update`.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCampaignRequest"
      responses:
        "200":
          description: Cập nhật thành công
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
        "403":
          description: Không có quyền, hoặc chiến dịch không ở trạng thái được sửa
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteCampaign
      summary: Xóa chiến dịch
      description: >-
        Chỉ chủ dự án hoặc admin, và chỉ khi chiến dịch ở `draft`, `rejected` hoặc `needs_info`
        (chưa từng công khai). Chiến dịch đã phát hành trả 409 — kết thúc qua luồng moderate.
        Ghi audit `campaign.delete`.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: Xóa thành công
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Chiến dịch đã phát hành, không được xóa
  /campaigns/{id}/submit:
    post:
      operationId: submitCampaign
      summary: Gửi chiến dịch để admin xét duyệt
      description: >-
        Chỉ khi đang ở `draft`, `rejected` hoặc `needs_info` → chuyển sang
        `pending`. Ghi audit `campaign.submit`.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: Đã gửi duyệt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
        "403":
          description: Không có quyền, hoặc chiến dịch không ở trạng thái gửi duyệt được
        "404":
          $ref: "#/components/responses/NotFound"
  /campaigns/{id}/moderate:
    patch:
      operationId: moderateCampaign
      summary: Admin duyệt/từ chối/tạm dừng/kết thúc chiến dịch
      description: >-
        Chỉ admin. Bắt buộc theo ma trận chuyển trạng thái (vd. không thể
        chuyển `draft` → `ended`); bắt buộc `reason` khi chuyển sang
        `rejected`, `needs_info` hoặc `paused`. Ghi audit `campaign.moderate`.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ModerateCampaignRequest"
      responses:
        "200":
          description: Đã cập nhật trạng thái
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
        "400":
          description: Thiếu `reason` bắt buộc
        "403":
          description: Không phải admin
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Chuyển trạng thái không hợp lệ theo vòng đời chiến dịch
...
  /ai/recommend:
    post:
      operationId: aiRecommend
      summary: Gợi ý chiến dịch
      description: >-
        Đăng nhập tùy chọn (`OptionalJwtAuthGuard`) — có JWT thì dùng lịch sử
        hành vi/tài trợ để cá nhân hóa. Luôn trả `available: true`: lỗi/timeout
        AI service sẽ tự rơi về `source: POPULAR_FALLBACK` (xếp theo độ phổ
        biến), không bao giờ trả lỗi cho client.
      security:
        - bearerAuth: []
        - {}
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecommendRequest"
      responses:
        "200":
          description: Danh sách gợi ý (có lý do ngắn cho mỗi gợi ý)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecommendResult"
  /ai/predict:
    post:
      operationId: aiPredict
      summary: Dự đoán khả năng thành công của một chiến dịch (chỉ admin)
      description: >-
        `data` chỉ có khi `available: true` (AI service phản hồi được);
        `false` khi AI service lỗi/timeout — không dùng làm căn cứ từ chối,
        chỉ hỗ trợ tham khảo. Chủ dự án dùng `GET /campaigns/{id}/prediction`.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PredictRequest"
      responses:
        "200":
          description: "Kết quả dự đoán (hoặc `available: false` nếu AI service lỗi)"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PredictResult"
        "401":
          description: Chưa đăng nhập
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: Chiến dịch không tồn tại
  /ai/fraud:
    post:
      operationId: aiFraud
      summary: Chấm điểm rủi ro gian lận cho chiến dịch hoặc tài khoản (chỉ admin)
      description: Phải cung cấp ít nhất `campaignId` hoặc `userId`.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FraudRequest"
      responses:
        "200":
          description: "Điểm rủi ro (hoặc `available: false` nếu AI service lỗi)"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FraudResult"
        "400":
          description: Thiếu cả `campaignId` lẫn `userId`
        "403":
          $ref: "#/components/responses/Forbidden"
  /ai/events:
    post:
      operationId: recordBehaviorEvent
      summary: Ghi sự kiện hành vi (xem/theo dõi/tài trợ) phục vụ cá nhân hóa
      description: >-
        Chỉ ghi khi người dùng đã đồng ý cá nhân hóa (`aiTrackingConsent`). Chưa đồng ý thì
        server bỏ qua và trả `{ recorded: false, reason: "NO_CONSENT" }`.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBehaviorEventRequest"
      responses:
        "201":
          description: Kết quả ghi nhận
          content:
            application/json:
              schema:
                type: object
                properties:
                  recorded:
                    type: boolean
                  eventId:
                    type: string
                    format: uuid
                  reason:
                    type: string
                    enum: [NO_CONSENT]
        "404":
          description: Chiến dịch không tồn tại
...
  /admin/categories:
    get:
      operationId: adminListCategories
      summary: Toàn bộ lĩnh vực kèm số chiến dịch (chỉ admin)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Danh sách lĩnh vực
          content:
            application/json:
              schema:
                type: array
                items:
                  allOf:
                    - $ref: "#/components/schemas/Category"
                    - type: object
                      properties:
                        campaignCount: { type: integer }
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      operationId: adminCreateCategory
      summary: Thêm lĩnh vực (chỉ admin)
      description: Tên không trùng (không phân biệt hoa thường). Audit `category.create`.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string, minLength: 2, maxLength: 100 }
                description: { type: string, maxLength: 300 }
                sortOrder: { type: integer, minimum: 0, maximum: 1000 }
      responses:
        "201":
          description: Đã tạo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          description: Trùng tên
  /admin/categories/{id}:
    patch:
      operationId: adminUpdateCategory
      summary: Đổi tên/sắp xếp/tắt-bật lĩnh vực (chỉ admin)
      description: >-
        Đổi tên cập nhật luôn `campaigns.category` trong cùng giao dịch. Tắt: không nhận
        chiến dịch mới (POST/PATCH /campaigns trả 400), chiến dịch cũ giữ nguyên; phải còn
        ít nhất một lĩnh vực đang nhận. Audit `category.update`.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string, minLength: 2, maxLength: 100 }
                description: { type: string, maxLength: 300 }
                sortOrder: { type: integer, minimum: 0, maximum: 1000 }
                isActive: { type: boolean }
      responses:
        "200":
          description: Đã cập nhật
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Trùng tên hoặc tắt lĩnh vực cuối cùng
...
  /campaigns/{id}/stats:
    get:
      operationId: getCampaignStats
      summary: Số liệu chiến dịch cho chủ dự án/admin
      description: >-
        Lượt xem, lượt ủng hộ, người theo dõi, tỷ lệ chuyển đổi (lượt ủng hộ thành công /
        lượt xem) và chuỗi N ngày UTC gần nhất (lượt xem, số giao dịch đã xác nhận, số tiền
        còn giữ — giao dịch đã hoàn tiền không cộng tiền).
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdPath"
        - name: days
          in: query
          schema:
            type: integer
            minimum: 7
            maximum: 90
            default: 30
      responses:
        "200":
          description: Số liệu
          content:
            application/json:
              schema:
                type: object
                properties:
                  campaignId: { type: string, format: uuid }
                  viewCount: { type: integer }
                  backerCount: { type: integer }
                  followerCount: { type: integer }
                  currentAmount: { type: number }
                  goalAmount: { type: number }
                  conversionRate: { type: number, nullable: true, description: "0–1; null khi chưa có lượt xem" }
                  days: { type: integer }
                  daily:
                    type: array
                    items:
                      type: object
                      properties:
                        date: { type: string, format: date }
                        views: { type: integer }
                        donations: { type: integer }
                        amount: { type: number }
        "401":
          description: Chưa đăng nhập
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /campaigns/{id}/prediction:
    get:
      operationId: getCampaignPrediction
      summary: Ước lượng khả năng đạt mục tiêu (chỉ chủ dự án/admin)
      description: >-
        Không hiển thị công khai để tránh ảnh hưởng quyết định của người ủng hộ.
        Chỉ để tham khảo, không dùng làm căn cứ duyệt/từ chối. `available: false`
        khi AI service lỗi/timeout. Rate limit 30 yêu cầu/phút.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: "Kết quả dự đoán (hoặc `available: false` nếu AI service lỗi)"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PredictResult"
        "401":
          description: Chưa đăng nhập
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          description: Vượt giới hạn tần suất
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    Offset:
      name: offset
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    CampaignQ:
      name: q
      in: query
      description: Tìm kiếm toàn văn (tiêu đề/mô tả/địa điểm hoặc tên/email tùy endpoint)
      schema:
        type: string
        maxLength: 200
    CampaignCategory:
      name: category
      in: query
      schema:
        type: string
        maxLength: 100
    CampaignStatusFilter:
      name: status
      in: query
      description: Một hoặc nhiều trạng thái, cách nhau bởi dấu phẩy (vd. `active,success`)
      style: form
      explode: false
      schema:
        type: array
        items:
          $ref: "#/components/schemas/CampaignStatus"
    CampaignSort:
      name: sort
      in: query
      schema:
        type: string
        enum: [featured, popular, ending, newest, progress, latest]
        default: popular
      description: "`featured`: dự án admin chọn nổi bật trước (mới chọn trước), còn lại theo độ phổ biến"
  responses:
    Forbidden:
      description: Không có quyền thực hiện hành động này
    NotFound:
      description: Không tìm thấy tài nguyên
...
  schemas:
...
    UserSummary:
      type: object
      required: [id, email, name]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [user, campaign_owner, admin]
        emailVerified:
          type: boolean
          description: Có trong phản hồi đăng nhập/đăng ký/refresh/đổi mật khẩu
        aiTrackingConsent:
          type: boolean
          description: Đã đồng ý ghi nhận hành vi cho gợi ý AI
        aiConsentDecided:
          type: boolean
          description: false = chưa từng chọn → frontend hiển thị lời mời đồng ý
...
    PageInfo:
      type: object
      required: [total, limit, offset]
      properties:
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
    CampaignStatus:
      type: string
      enum:
        - draft
        - pending
        - approved
        - rejected
        - needs_info
        - active
        - paused
        - success
        - failed
        - cancelled
        - ended
    Category:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        description: { type: string, nullable: true }
        sortOrder: { type: integer }
        isActive: { type: boolean }
    Campaign:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        category:
          type: string
        ownerId:
          type: string
          format: uuid
        owner:
          nullable: true
          $ref: "#/components/schemas/UserSummary"
        goalAmount:
          type: number
        currentAmount:
          type: number
          description: Sinh từ tổng giao dịch `completed`, không nhận nhập tay
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        status:
          $ref: "#/components/schemas/CampaignStatus"
        isFeatured:
          type: boolean
          description: Admin chọn hiển thị ở mục "Dự án nổi bật" trang chủ
        imageUrl:
          type: string
          nullable: true
        videoUrl:
          type: string
          nullable: true
        location:
          type: string
          nullable: true
        backerCount:
          type: integer
        viewCount:
          type: integer
        rejectionReason:
          type: string
          nullable: true
          description: Dùng chung cho lý do từ chối, yêu cầu bổ sung hoặc tạm dừng
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateCampaignRequest:
      type: object
      required: [title, description, category, goalAmount, endDate]
      properties:
        title:
          type: string
          maxLength: 200
        description:
          type: string
        category:
          type: string
          maxLength: 100
        goalAmount:
          type: number
          minimum: 1000
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        imageUrl:
          type: string
          format: uri
        location:
          type: string
          maxLength: 255
    UpdateCampaignRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 200
        description:
          type: string
        category:
          type: string
          maxLength: 100
        goalAmount:
          type: number
          minimum: 1000
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        imageUrl:
          type: string
          format: uri
        location:
          type: string
          maxLength: 255
    ModerateCampaignRequest:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [approved, active, paused, rejected, needs_info, ended]
        reason:
          type: string
          description: Bắt buộc khi `status` là `rejected`, `needs_info` hoặc `paused`
    CampaignList:
      allOf:
        - $ref: "#/components/schemas/PageInfo"
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/Campaign"
...
    RecommendRequest:
      type: object
      properties:
        preferences:
          type: array
          maxItems: 10
          items:
            type: string
            maxLength: 100
        excludeIds:
          type: array
          maxItems: 50
          items:
            type: string
            format: uuid
        limit:
          type: integer
          minimum: 1
          maximum: 20
    RecommendResult:
      type: object
      required: [available, items]
      properties:
        available:
          type: boolean
        source:
          type: string
          enum: [COLLABORATIVE, CONTENT, COLD_START, POPULAR_FALLBACK]
        items:
          type: array
          items:
            type: object
            required: [campaign, score, reason]
            properties:
              campaign:
                $ref: "#/components/schemas/Campaign"
              score:
                type: number
              reason:
                type: string
                description: Lý do gợi ý ngắn, hiển thị trực tiếp cho người dùng
        detail:
          type: string
    PredictRequest:
      type: object
      required: [campaignId]
      properties:
        campaignId:
          type: string
          format: uuid
    AiPredictFactor:
      type: object
      properties:
        feature:
          type: string
        label:
          type: string
        direction:
          type: string
          enum: [UP, DOWN]
        weight:
          type: number
        note:
          type: string
    PredictResult:
      type: object
      required: [available, campaignId]
      properties:
        available:
          type: boolean
        campaignId:
          type: string
          format: uuid
        data:
          nullable: true
          type: object
          description: Chỉ có khi `available` là `true`
          properties:
            probability:
              type: number
            prediction:
              type: string
              enum: [LIKELY, UNLIKELY]
            confidence:
              type: number
            model:
              type: string
              nullable: true
            fallback:
              type: boolean
            metrics:
              type: object
              nullable: true
              additionalProperties: true
            factors:
              type: array
              items:
                $ref: "#/components/schemas/AiPredictFactor"
    FraudRequest:
      type: object
      description: Cần ít nhất một trong `campaignId` hoặc `userId`
      properties:
        campaignId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
    FraudResult:
      type: object
      required: [available, entityType, entityId]
      properties:
        available:
          type: boolean
        entityType:
          type: string
          enum: [CAMPAIGN, USER]
        entityId:
          type: string
          format: uuid
        data:
          nullable: true
          type: object
          description: Chỉ có khi `available` là `true`
          properties:
            riskScore:
              type: number
            level:
              type: string
              enum: [HIGH, MEDIUM, LOW]
            method:
              type: string
              enum: [ENSEMBLE, RULE]
            reasons:
              type: array
              items:
                $ref: "#/components/schemas/RiskReason"
            evidences:
              type: object
              additionalProperties:
                type: number
            fallback:
              type: boolean
    CreateBehaviorEventRequest:
      type: object
      required: [campaignId, eventType]
      properties:
        campaignId:
          type: string
          format: uuid
        eventType:
          type: string
          enum: [view, follow, contribute]
...
    RiskReason:
      type: object
      properties:
        group:
          type: string
        label:
          type: string
        weight:
          type: number
...